Add files via upload
Intended to use for Tenants that does not have a licens to get good enough report from Intune whether win10 devices are upgradeable to win11.

Haven't tried from Intune yet, but I don't see a reason why it wouldn't work as long as it is run in system context. The intent is to move the file to an isolated network folder that aren't mapped for the user but has no security to avoid 401 errors. Security by obfuscation, or something.